fix: CodeRabbit 리뷰 반영
- completeOrderPayment에 소유권 검증 + UNPAID 상태 가드 추가
- 스케줄러 조회를 Slice 기반 배치 처리로 변경 (BATCH_SIZE=100)
- Long 비교 != → longValue() 사용
- 중복 테스트 제거

+++ src/main/java/com/example/pdelivery/payment/application/PaymentServiceImpl.java
@@ -91,7 +91,7 @@ public class PaymentServiceImpl implements PaymentService {
 	public boolean approvePaymentByOrder(UUID orderId, Long amount) {
 		PaymentOrderSummary summary = paymentOrderRequirer.getOrderSummary(orderId);
 
-		if (amount == null || summary.totalAmount() != amount) {
+		if (amount == null || summary.totalAmount() != amount.longValue()) {
 			throw new PaymentException(PaymentErrorCode.INVALID_AMOUNT);
 		}
